Split syntax extensions into separate files
